feat: Move classes to authentication-ldap
